refactor: refactor modules, fix lint errors
