+++ .eslintrc.json
@@ -18,6 +18,6 @@
     "import/prefer-default-export": "off",
     "import/no-default-export": "error",
     "class-methods-use-this": "off",
-    "radix": "off"
+    "no-restricted-syntax": "off"
   }
 }
