+++ src/controller/user/createUserAction.ts
@@ -1,5 +1,5 @@
 import { constants as httpConstants } from 'node:http2';
-import { ActionHandler } from '../../http/ActionHandler';
+import { ActionHandler } from '../../types/ActionHandler';
 import { Request } from '../../http/Request';
 import { Response } from '../../http/Response';
 import { getRequestBody } from '../../http/requestService';
