+++ src/types/Endpoint.ts
