+++ src/types/WorkerMessage.ts
@@ -1,0 +1,13 @@
+import { IUser } from '../components/users/user.interface';
+
+type WorkerMessageData = {
+    userId?: string;
+    user?: IUser;
+};
+
+type WorkerMessage = {
+    method: string;
+    parameters?: WorkerMessageData;
+};
+
+export { WorkerMessage };
